ms_bin: read: change the type of the count to off_t

+++ libms.h
@@ -50,7 +50,7 @@
 
 	typedef struct {
 		int fd;
-		int count, total;
+		off_t count, total;
 		int elem_size;
 		void (*bin_conv_b2h)(int*, void*, ms_state_t*);
 		struct bin_buf buf;
@@ -105,7 +105,7 @@
 	void ms_bin_seq_read_open(const char *filename, ms_bin_seq_read_flag_t flag, ms_bin_seq_read_t *mbp, ms_state_t *st);
 	void ms_bin_seq_read_close(ms_bin_seq_read_t *mbp, ms_state_t *st);
 	ms_bin_ret_t ms_bin_seq_read_next(int *ms, ms_bin_seq_read_t *mbp, ms_state_t *st);
-	void ms_bin_seq_read_seek(int count, int whence, ms_bin_seq_read_t *mbp, ms_state_t *st);
+	void ms_bin_seq_read_seek(off_t count, int whence, ms_bin_seq_read_t *mbp, ms_state_t *st);
 	void ms_bin_seq_write_open(const char *filename, ms_bin_seq_write_flag_t flag, ms_bin_seq_write_t *mbp, ms_state_t *st);
 	void ms_bin_seq_write_close(ms_bin_seq_write_t *mbp, ms_state_t *st);
 	void ms_bin_seq_write_next(int *ms, ms_bin_seq_write_t *mbp, ms_state_t *st);
